# src/main/resources/com/sonicle/webtop/contacts/openapi-v2.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "WebTop Contacts",
    "contact": {
      "email": "dev-team@sonicle.com"
    },
    "version": "v2",
    "description": "This is Contacts service API enpoint."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "auth-basic": []
    },
    {
      "auth-bearer": []
    }
  ],
  "paths": {
...
    "/eas/folders/{folderId}/messages-stats": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "List all messages for a specific folder",
        "operationId": "getMessagesStats",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EasSyncContactStat"
                  }
                }
              }
            }
          }
        },
        "description": "Returns sync informations for the specified Category."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages": {
      "post": {
        "tags": [
          "eas"
        ],
        "summary": "Adds a message",
        "operationId": "addMessage",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncContactUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncContactStat"
                }
              }
            }
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Add new Contact into specified Category."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        }
      ]
    },
    "/eas/folders/{folderId}/messages/{id}": {
      "get": {
        "tags": [
          "eas"
        ],
        "summary": "Get a single message",
        "operationId": "getMessage",
        "parameters": [
          {
            "name": "picture",
            "in": "query",
            "description": "Determine whether to return picture data",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncContact"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Contact not found",
            "content": {}
          }
        },
        "description": "Gets the specified Contact."
      },
      "put": {
        "tags": [
          "eas"
        ],
        "summary": "Updates a message",
        "operationId": "updateMessage",
        "parameters": [
          {
            "name": "picture",
            "in": "query",
            "description": "Determine whether to update picture data",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EasSyncContactUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EasSyncContactStat"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Contact not found",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body",
        "description": "Updates the specified Contact."
      },
      "delete": {
        "tags": [
          "eas"
        ],
        "summary": "Deletes a message",
        "operationId": "deleteMessage",
        "responses": {
          "204": {
            "description": "Success",
            "content": {}
          },
          "400": {
            "description": "Invalid parameter",
            "content": {}
          },
          "404": {
            "description": "Contact not found",
            "content": {}
          }
        },
        "description": "Deletes the specified Contact."
      },
      "parameters": [
        {
          "name": "folderId",
          "in": "path",
          "description": "Folder ID",
          "required": true,
          "schema": {
            "type": "string"
          },
          "example": "1234"
        },
        {
          "name": "id",
          "in": "path",
          "description": "Message ID",
          "required": true,
          "schema": {
            "type": "string"
          }
        }
      ]
    },
...
    "/me/contacts/{contact_id}/picture": {
      "parameters": [
        {
          "schema": {
            "type": "string"
          },
          "name": "contact_id",
          "in": "path",
          "required": true,
          "description": "The ID of a contact"
        }
      ],
      "get": {
        "summary": "Get a contact picture metadata",
        "tags": [
          "me"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContactPictureMeta"
                }
              }
            }
          },
          "404": {
            "description": "Contact not found",
            "content": {}
          }
        },
        "operationId": "getContactPictureMeta",
        "description": "Gets the contact's picture metatada of a given contact ID."
      },
      "put": {
        "summary": "Set contact picture",
        "tags": [
          "me"
        ],
        "operationId": "setContactPicture",
        "responses": {
          "200": {
            "description": "Success"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "description": "Assign a picture to the given contact ID. The picture should be in binary. It replaces any existing photo for that contact.",
        "requestBody": {
          "content": {
            "image/*": {
              "schema": {
                "type": "string",
                "format": "binary"
              }
            }
          },
          "description": "The binary data of the picture."
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "header",
            "name": "Content-Type",
            "description": "The picture media-type."
          }
        ]
      },
      "delete": {
        "summary": "",
        "operationId": "clearContactPicture",
        "responses": {
          "204": {
            "description": "Contact's picture deleted"
          },
          "404": {
            "description": "Contact not found"
          }
        },
        "description": "Deletes the contact's picture of a given contact ID.",
        "tags": [
          "me"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EasSyncContactStat": {
        "type": "object",
        "description": "Bean for carry messageStat fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          }
        },
        "required": [
          "id",
          "etag"
        ]
      },
      "EasSyncContact": {
        "type": "object",
        "description": "Bean for carry messageStat fields",
        "properties": {
          "id": {
            "type": "string",
            "description": "Message ID (internal)"
          },
          "etag": {
            "type": "string",
            "description": "Revision tag"
          },
          "title": {
            "type": "string",
            "description": "Title"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "nickname": {
            "type": "string",
            "description": "Nickname"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile phone"
          },
          "pager1": {
            "type": "string",
            "description": "Pager 1"
          },
          "pager2": {
            "type": "string",
            "description": "Pager 2"
          },
          "email1": {
            "type": "string",
            "description": "Email address 1"
          },
          "email2": {
            "type": "string",
            "description": "Email address 2"
          },
          "email3": {
            "type": "string",
            "description": "Email address 3"
          },
          "im1": {
            "type": "string",
            "description": "Instant Msg. 1"
          },
          "im2": {
            "type": "string",
            "description": "Instant Msg. 2"
          },
          "im3": {
            "type": "string",
            "description": "Instant Msg. 3"
          },
          "workAddress": {
            "type": "string",
            "description": "Business address"
          },
          "workPostalCode": {
            "type": "string",
            "description": "Business postal code"
          },
          "workCity": {
            "type": "string",
            "description": "Business city"
          },
          "workState": {
            "type": "string",
            "description": "Business state/province"
          },
          "workCountry": {
            "type": "string",
            "description": "Business country"
          },
          "workTelephone1": {
            "type": "string",
            "description": "Business telephone 1"
          },
          "workTelephone2": {
            "type": "string",
            "description": "Business telephone 2"
          },
          "workFax": {
            "type": "string",
            "description": "Business fax"
          },
          "homeAddress": {
            "type": "string",
            "description": "Home address"
          },
          "homePostalCode": {
            "type": "string",
            "description": "Home postal code"
          },
          "homeCity": {
            "type": "string",
            "description": "Home city"
          },
          "homeState": {
            "type": "string",
            "description": "Home state"
          },
          "homeCountry": {
            "type": "string",
            "description": "Home country"
          },
          "homeTelephone1": {
            "type": "string",
            "description": "Home telephone 1"
          },
          "homeTelephone2": {
            "type": "string",
            "description": "Home telephone 2"
          },
          "homeFax": {
            "type": "string",
            "description": "Home fax"
          },
          "otherAddress": {
            "type": "string",
            "description": "Other address"
          },
          "otherPostalCode": {
            "type": "string",
            "description": "Other postal code"
          },
          "otherCity": {
            "type": "string",
            "description": "Other city"
          },
          "otherState": {
            "type": "string",
            "description": "Other state"
          },
          "otherCountry": {
            "type": "string",
            "description": "Other country"
          },
          "companyId": {
            "type": "string",
            "description": "Company ID"
          },
          "companyName": {
            "type": "string",
            "description": "Company name"
          },
          "function": {
            "type": "string",
            "description": "Job title"
          },
          "department": {
            "type": "string",
            "description": "Department"
          },
          "manager": {
            "type": "string",
            "description": "Manager name"
          },
          "assistant": {
            "type": "string",
            "description": "Assistant name"
          },
          "assistantTelephone": {
            "type": "string",
            "description": "Assistant telephone"
          },
          "partner": {
            "type": "string",
            "description": "Partner"
          },
          "birthday": {
            "type": "string",
            "description": "Birthday date (ISO date YYYYMMDD)"
          },
          "anniversary": {
            "type": "string",
            "description": "Anniversary date (ISO date YYYYMMDD)"
          },
          "url": {
            "type": "string",
            "description": "Web-site URL"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          },
          "picture": {
            "type": "string",
            "description": "Image base64 data URI"
          }
        },
        "required": [
          "id",
          "etag"
        ]
      },
      "EasSyncContactUpdate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Title"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "nickname": {
            "type": "string",
            "description": "Nickname"
          },
          "mobile": {
            "type": "string",
            "description": "Mobile phone"
          },
          "pager1": {
            "type": "string",
            "description": "Pager 1"
          },
          "pager2": {
            "type": "string",
            "description": "Pager 2"
          },
          "email1": {
            "type": "string",
            "description": "Email address 1"
          },
          "email2": {
            "type": "string",
            "description": "Email address 2"
          },
          "email3": {
            "type": "string",
            "description": "Email address 3"
          },
          "im1": {
            "type": "string",
            "description": "Instant Msg. 1"
          },
          "im2": {
            "type": "string",
            "description": "Instant Msg. 2"
          },
          "im3": {
            "type": "string",
            "description": "Instant Msg. 3"
          },
          "workAddress": {
            "type": "string",
            "description": "Business address"
          },
          "workPostalCode": {
            "type": "string",
            "description": "Business postal code"
          },
          "workCity": {
            "type": "string",
            "description": "Business city"
          },
          "workState": {
            "type": "string",
            "description": "Business state/province"
          },
          "workCountry": {
            "type": "string",
            "description": "Business country"
          },
          "workTelephone1": {
            "type": "string",
            "description": "Business telephone 1"
          },
          "workTelephone2": {
            "type": "string",
            "description": "Business telephone 2"
          },
          "workFax": {
            "type": "string",
            "description": "Business fax"
          },
          "homeAddress": {
            "type": "string",
            "description": "Home address"
          },
          "homePostalCode": {
            "type": "string",
            "description": "Home postal code"
          },
          "homeCity": {
            "type": "string",
            "description": "Home city"
          },
          "homeState": {
            "type": "string",
            "description": "Home state"
          },
          "homeCountry": {
            "type": "string",
            "description": "Home country"
          },
          "homeTelephone1": {
            "type": "string",
            "description": "Home telephone 1"
          },
          "homeTelephone2": {
            "type": "string",
            "description": "Home telephone 2"
          },
          "homeFax": {
            "type": "string",
            "description": "Home fax"
          },
          "otherAddress": {
            "type": "string",
            "description": "Other address"
          },
          "otherPostalCode": {
            "type": "string",
            "description": "Other postal code"
          },
          "otherCity": {
            "type": "string",
            "description": "Other city"
          },
          "otherState": {
            "type": "string",
            "description": "Other state"
          },
          "otherCountry": {
            "type": "string",
            "description": "Other country"
          },
          "companyId": {
            "type": "string",
            "description": "Company ID"
          },
          "companyName": {
            "type": "string",
            "description": "Company name"
          },
          "function": {
            "type": "string",
            "description": "Job title"
          },
          "department": {
            "type": "string",
            "description": "Department"
          },
          "manager": {
            "type": "string",
            "description": "Manager name"
          },
          "assistant": {
            "type": "string",
            "description": "Assistant name"
          },
          "assistantTelephone": {
            "type": "string",
            "description": "Assistant telephone"
          },
          "partner": {
            "type": "string",
            "description": "Partner"
          },
          "birthday": {
            "type": "string",
            "description": "Birthday date (ISO date YYYYMMDD)"
          },
          "anniversary": {
            "type": "string",
            "description": "Anniversary date (ISO date YYYYMMDD)"
          },
          "url": {
            "type": "string",
            "description": "Web-site URL"
          },
          "notes": {
            "type": "string",
            "description": "Notes"
          },
          "picture": {
            "type": "string",
            "description": "Image base64 data URI"
          }
        },
        "description": "Bean for carry message's updateable fields"
      },
...
      "ContactPictureMeta": {
        "title": "ContactPictureMeta",
        "x-stoplight": {
          "id": "myzl3iqcsmwwt"
        },
        "type": "object",
        "description": "Represent data of a contact's picture.",
        "properties": {
          "mediaType": {
            "type": "string",
            "x-stoplight": {
              "id": "yr7psum2uzzeo"
            },
            "description": "The image MediaType: like 'image-<subtype>'",
            "readOnly": true
          },
          "width": {
            "type": "integer",
            "x-stoplight": {
              "id": "q0vi63nkj9os5"
            },
            "description": "The image width in pixels.",
            "readOnly": true
          },
          "height": {
            "type": "integer",
            "x-stoplight": {
              "id": "1maok9niii9kv"
            },
            "description": "The image height in pixels.",
            "readOnly": true
          },
          "size": {
            "type": "integer",
            "x-stoplight": {
              "id": "d7aggjpfulfo8"
            },
            "description": "The image size in bytes.",
            "format": "int64",
            "readOnly": true
          }
        },
        "required": [
          "mediaType"
        ]
      },
...
    },
    "securitySchemes": {
      "auth-basic": {
        "type": "http",
        "scheme": "basic"
      },
      "auth-bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
  },
...
}